% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotool.R
\name{load_fasta_file}
\alias{load_fasta_file}
\title{Load a FASTA file}
\usage{
load_fasta_file(filename, verbose = FALSE)
}
\arguments{
\item{filename}{The name of the input FASTA file.}

\item{verbose}{A logical value indicating whether the function is verbose}
}
\value{
A character vector of containing sequenecs as characters. Each
    element in the vector is a sequence in the FASTA file.
}
\description{
Wrapper for seqinr \code{read.fasta()} to load in a FASTA file
    and get the sequences.
}
\examples{
my_fasta_file <- system.file("extdata", "two_sequence.fasta", package = "biotool")
load_fasta_file(filename = my_fasta_file)
}

