% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsea2}
\alias{plot_gsea2}
\title{Plot Gene Set Enrichment Analysis (GSEA) Results}
\usage{
plot_gsea2(gsea_, term_)
}
\arguments{
\item{gsea_}{A GSEA result object from the `clusterProfiler` package.}

\item{term_}{The term to visualize.}
}
\value{
A ggplot object representing the GSEA plot.
}
\description{
Generates a GSEA plot for a specific term.
}
