% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP visualization}
\usage{
plot_umap(
  df_,
  colorBy_ = "seurat_clusters",
  x_ = "umap_1",
  y_ = "umap_2",
  label_ = T,
  border_ = T,
  arrowSize_ = 0.5
)
}
\arguments{
\item{df_}{Data frame. Contains UMAP coordinates and grouping information.}

\item{colorBy_}{Character. The column name for the grouping variable used for coloring points.}

\item{x_}{Character. The column name for the UMAP x-axis.}

\item{y_}{Character. The column name for the UMAP y-axis.}

\item{label_}{Logic. Whether to add label.}

\item{border_}{Logic. Whether to add border.}

\item{arrowSize_}{Numeric. Relative size of arrow.}
}
\value{
A ggplot object displaying the UMAP visualization.
}
\description{
Creates a UMAP plot with points colored by the specified grouping variable.
}
\examples{
plot_umap(df_, x_ = "UMAP_1", y_ = "UMAP_2", colorBy_ = "CellType")

}
