% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotool.R
\name{pretty_output}
\alias{pretty_output}
\title{Prettify output}
\usage{
pretty_output(stats)
}
\arguments{
\item{stats}{The list containing FASTA stats.}
}
\value{
A list containing FASTA stats
}
\description{
Use a dash (-) in place of min, avg, and max if numseq is zero.
}
\examples{
my_fasta_file <- system.file("extdata", "empty_file", package = "biotool")
sequences <- load_fasta_file(filename = my_fasta_file)
stats <- get_seq_stats(sequences, filename = my_fasta_file)
pretty_output(stats)
}

