% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotool.R
\name{run_biotool}
\alias{run_biotool}
\title{Biotool}
\usage{
run_biotool(fasta_files, min_len = 0, pretty = TRUE, verbose = FALSE,
  quit_on_error = TRUE)
}
\arguments{
\item{fasta_files}{Vector of filenames}

\item{min_len}{The minimum length of the sequence to include when
calculating statistics}

\item{pretty}{A dash (-) is used in place of min, avg, and max if numseq is
zero.}

\item{verbose}{A logical value indicating whether the function is verbose}

\item{quit_on_error}{A logical value indicating if an error when reading a
fasta file will invoke the \code{quit} function to terminate R with an
exit status of 3. This will only occur if the R session is
non-interactive.}
}
\value{
A data frame with FASTA stats. Each row represents a FASTA file.
}
\description{
For the given fasta files, calculate statistics and return
    results in a data frame.
}
\examples{
my_fasta_files <- c(
    system.file("extdata", "one_sequence.fasta", package = "biotool"),
    system.file("extdata", "two_sequence.fasta", package = "biotool"))
run_biotool(my_fasta_files)
}

